BUGFIX - translate 'core' => 'free' for "whats new"
While QA-ing the "whats new" feature on staging, I noticed that there
was a bug in translating "free". I've updated the logic so it correctly
translates "core" to "free".

I also titlecased the package names in the data yaml files. As well as
swapped out "All" for a list of package names.

The issue found here would be resolved once the file linter is merged:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/48851

But as that is still under development I figured a small MR to fix the
problem is best for now.