Update usages of files_max_capacity
Because there are now 4 models replicated by SSF. This is still a rough
stopgap, but it does not have to be 100% accurate since the behavior of
the system with different capacity numbers depends on the implementation
details as well as the particular characteristics of the deployment. The
main thing we need to resolve soon is that snippet repositories syncs
are limited by "files max capacity" instead of "repositories max
capacity". But that change is probably not as trivial as this one, so I
have opened a follow up issue:
https://gitlab.com/gitlab-org/gitlab/-/issues/294074