Move database spec helpers into separate directory
We'd like to run CI pipelines against officially supported PG versions
to ensure we're compatible with all of them. Since this isn't practical
to do for every MR created, we'd like to trigger those tests only when
specific files are edited that are likely to work differently across
versions.

We have a number of spec helpers for database, so this moves them all
into their own directory where it's easy to detect a file change when
the pipeline runs.