Change the interface of Security artifact parsers
Previously we were instantiating the parser objects without passing any
initialization arguments. Because of this, we were passing the report
and the JSON text to the parse method and all the other utility methods
which was causing inflation on method signatures.

By just passing the report and the JSON text as the initialization args
to the instances, we can avoid passing them in between utility methods.