Document Elasticsearch data recovery methods
Since we've always intended to have these various backup methods for
recovering data in our integration but we've never clearly documented
them we will write this down as a rough draft for now. If we are really
going to rely on this and encourage it's use we should also be testing
it and automating parts of it but that is more expensive and therefore
writing down what we know today is better than nothing.