PlantUML note about redirects when using TLS
The PlantUML service can be configured to serve traffic outside
of localhost, and the redirection can normally be avoided in
this situation.

However, on GitLab installations using TLS, new mixed-content browser
policies will perform a silent upgrade of HTTP connections to HTTPS
and/or refuse to load a HTTP resource altogether when TLS is in use.

The redirection becomes a requirement in this situation.

The change adds a note to the PlantUML redirect step
in docs to explain this.