Fix flaky spec in QuickActions::InterpretService#execute
The spec was attempting to test whether `/assign @userA @userB` would
only assign an issue or merge request to `userA`. However, this would
fail occasionally because the reference extractor does not guarantee
ordering with extracted Markdown references since a database lookup is
performed (e.g. `User.where`) without an ORDER BY.

We adapt the test to check that only one of the two possible users is
assigned since the behavior is ambiguous when multiple assignees is not
supported, and it would take some work with some performance
implications to ensure the first mentioned user were assigned.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/27989