Implement declarative enum abstraction
Instead of spreading the enum information all around the codebase
between different modules like models and GraphQL types, we can set a
single source of the truth for the Enum values.

To do so, we should promote the enums to be first class members of the
project by giving them a new top level directory to live in.