Avoid 409 StaleObjectError errors with /rebase
If a merge request is closed, the quick action /rebase will quietly fail
with a 409 StaleObjectError. To avoid that, we disable this quick action
if the merge request is closed.

If a rebase is in progress, the quick action will also fail with the
same error. We now check the state and report to the user that the
rebase is in progress.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/296223