Delete Elasticsearch migrations index in rake elastic:delete_index
The migrations Elasticsearch index was recently introduced but we didn't
add anything to our normal lifecycle rake tasks to delete it. This
should have been done initially and will be the expected user
experience.

This task is also used by `gitlab:elastic:index` which is the main way
for resetting your index back to scratch. It makes sense that these all
will clear out all index types including the migrations index. This will
help with any users trying to use this rake task to do a reset of
everything and will also help with local development.

This MR also makes 2 minor refactors to introduce methods
`#delete_migrations_index` and `#migrations_index_exists?` which should
simplify a lot of duplicated calls. Most of the changes in this MR is
refactoring to use these new helper methods.