This uses more sophisticated client query processing
Rather than skipping all queries that use the @client directive, we
instead remove all such fields (and any arguments and fragments
mentioned in the skipped sections) and then only skip the query if that
then leaves us with an empty query.

The query transformation is handled with a query printer.