Limit member tribute autocomplete to 10
We currently show all members in the autocomplete menu but since
each member menu item contains an <img> avatar, a large number
of fetches are done to get all these avatars. This commit limits
the number of members listed in the autocomplete menu to avoid
doing a large number of fetches.