Correctly handle Gitaly being unavailable in more locations
If `GRPC::Unavailable` or `Gitlab::Git::CommandError` is raised in code
outside of a controller action, such as an `around_action`, it can be
executed after a render call inside the controller has completed. When
this happens, ActionController detects it and raises a very unhelpful
`AbstractController::DoubleRenderError`, swallowing the original error.

This code was introduced as part of this MR:
https://gitlab.com/gitlab-org/gitlab-foss/-/merge_requests/114491

Circuit-breakers have since been removed, so it should be safe to
remove this code as well.