Remove ReadmeBlob since it isn't used
ReadmeBlob#rendered_markup isn't called, but intended to be
called by BlobViewer::Markup. ReadmeBlob uses SimpleDelegator
to delegate all missing methods to Blob, but when the simple_viewer
call bubbles up to Blob#simple_viewer, the class isn't ReadmeBlob
anymore, but Blob, so it doesn't respond to rendered_markup.

ReadmeBlob was only responsible for calling
Repository#rendered_markup that cached the content in Redis.
Currently, we cache every blob content in Redis, so we can
remove ReadmeBlob instead of fixing the above problem.