Geo Node Status 2.0 - Init
This change introduces the beginning
of the new UI for Geo Nodes Status.

This change is behind the :geo_nodes_beta
feature flag.

After our full rollout these files will take
over the current Geo Node Status view.

For now we plan to roll it out on a seperate route
that will for a time exist along side the normal
route before taking it over.

More can be read about this approach on the epic:
https://gitlab.com/groups/gitlab-org/-/epics/4712