Integrate GlModal close button change
GlModal was updated to use a Pajamas-compliant close button. This is
done by overriding the modal-header slot to display a custom close
button, and by re-exposing the title prop as well as the modal-title
slot.
Those are drop-in changes as it still possible to set a modal's title
either by providing a modal-header slot, a modal-title slot, or by
passing the title prop.
However, the specs needed to be updated to reflect those API changes.