Add migration helper to create table with CHECKs
Add a new migration helper that allows the caller to define CHECK
constraints in the table definition block. This can be used to avoid
having to `disable_ddl_transaction!` and add the check constraints using
the existing helper, which cannot be used inside a transaction block.
Also provides a convenience method to more easily add CHECK constraints
for text limits.