Allow blank S3 regions to be used
If consolidated object storage is used without a region, Workhorse will
fall back to making a direct HTTP PUT call to the S3 endpoint because
GitLab Rails does not send the multipart upload URLs in this mode.

However, this will result a 411 Length Required error because multipart
upload URLs must be used when the Content-Length header is not
available.

An S3 region is not strictly required, particularly with Minio. With
Amazon S3, the region will default to US East (`us-east-1`). If the
bucket does not exist in the region, S3 will return with a more
informative error.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/297227