Move Ci pipeline graph mounted logic to an immediate watcher
Because pipeline_editor component can be used in multiple sections,
we cannot know if the rendering will only happen in mount lifecycle.
Instead, we move the logic to an immediate watcher where the render
can be trigger if:
- The component mounts for the first time
- The pipelineData prop changes

This also merges other watchers that were made to report error since
they are also dependants on the pipelineData.