Revert "Move help_page_helper to lib/utils"
This reverts commit 9a80dadaa797b6f11b125507f18a3af2789b6064.

A util is something which works outside of the project (i.e. could be
ported to another project without relying on any project specific code)
while a helper is something which augments or assists project specific
code. Like a helper which uses a class variable to augment the value.

Since this is specifically about generating help links for the GitLab
product, this therefore is a helper.