Enable multipart transfers for S3 copying
This is a follow-up to
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/50922. Multithreaded,
multipart transfers are only enabled if the file is above 5 GB or the
multipart chunk size is explicitly set.

We now set the chunk size to 10 megabytes (the AWS SDK default) to make
file copies faster.

This commit also updates fog-aws to v3.8.0 to work properly with server
side encryption parameters.

CHANGELOG: https://github.com/fog/fog-aws/blob/master/CHANGELOG.md