Limit tribute autocomplete menu items to 100
When testing tribute autocomplete on gitlab.com, issue
autocomplete renders tens of thousands of items in the
autocomplete menu which freezes the UI for a couple of
seconds. This is because with tribute it renders all
list items by default. To improve performance, this
commit adds a limit of 100 items, which is still more
than the limit of 5 we currently have with atjs
autocomplete.