Remove vue/order-in-components override
The rule override was in place to make the transition to the latest
`vue/recommended` ruleset easier, i.e., tackled iteratively.

Now that most violations of this rule have been fixed, this override can
be removed, so that we adopt the stricter version from
`vue/recommended`.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/297216.