Use old purchase flow when user last name is blank
The customers app requires a last name for the creation of the customer
record. As the new purchase flow does not currently have a way for users
to set a last name, fallback to the old flow when the user does not have
a last name set. This was causing issues with users in the following
state:

  1. User does not have a last name (signed up before we split the name
     field) and their full name only has one word
  2. User does not have an existing customers dot account
  3. User is creating the subscription on .com
  4. User is creating the subscription for a group (ie, using the
     purchase flow that does not redirect to customers-dot)

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/298721