Refactor Promote Milestone Modal
This refactors the "Promote Milestone Modal", replacing DeprecatedModal2
with GlModal. The way the modal is triggered, is changed from an
event-bus based solution to one which is "pure" Vue, simplifying the
code quite a bit. We moved most of the logic from the init function over
to the modal itself. Tests have been adjusted.