Move jest spec helpers to own directory
It is a bit confusing as we have put helpers to be used in specs in the
same directory as specs for `app/assets/javascripts/helpers`. In oder
to make the destinction a bit more clear, this moves jest test helpers
to `__helpers__` and adds appropriate readmes to both folders.