Use a dynamic loader for window.gapi
This MR swaps out an implicit dependency on window.gapi for a lazy
loader using a script tag designed around the realities of how
gapi needs to be loaded.

Previously, we had an implicit global dependency on gapi, relying on a
DOMContentLoaded event listener in an entirely different module. This
was brittle and held back internal efforts to remove dependencies on the
DOMContentLoaded event.

See https://gitlab.com/gitlab-org/gitlab/-/issues/284997