Move `detailed_description` to `help`
Changes apply only to confluence and github.

Changing this content to `help` means that the UI
is consistent with all other integrations (the
`help` content is displayed in the right-hand
column of the integration setup form)

Also, Ensures that Confluence and Github's "help"
doesn't interfere when trying to figure them at the
group or instance level.