Filter out duplicates from the finder by related commit sha
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/21686

**Problem**

We discovered an edge case when the same merge request can be returned
multiple times. If the project has `always squash` option on, then we
create a squash commit even for merge requests with only one
commit. In this case, the commit from the merge request is a squash
commit at the same time.

**Solution**

Use UNION instead of UNION ALL to remove duplicated records.