Add default user permissions const for Web IDE
This constant is what's returned by our
getter when the project data isn't available
yet.

This way we can prevent showing an alert
based on insufficient permissions until
we've actually fetched the permissions. Using
a default value, helps us be future-proof
so we don't need to add more conditions everywhere.