Add merge_request_refs_cleanup feature flag
We are experiencing performance issues with the merge request
refs cleanup functionality on production and some customers
instances. This became more visible when we started backfilling
`MergeRequest::CleanupSchedule` records for old merge requests.

Disabling the worker and scheduler for now behind a feature flag
so we can test the feature once we improve it's scheduling logic
before enabling it on other instances.