Refactor IssuableBaseService label processing
In issuable_base_service.rb the code processing labels is making
the update() method not as clear as it could be. Let's refactor
this code into a new assign_requested_labels() method.

While at it, lets rename labels_changing? into ids_changing?
as the method is not specific to labels and could be used with
other arrays of ids.