testing: Enable all Gitaly feature flags in CI
Gitaly feature flags are by default turned off for this test suite,
unless explicitly defaulting to on as set by Gitaly developers. This
creates a situation where almost always all feature flags are turned off
for Gitaly, and this test suite never runs fully with newer features.
Than when the flag is flipped on production, we might catch bugs that
the testsuite of gitlab.git would've caught.

This change depends on a Gitaly change[1] that overrides the defaults to
always enable all features.

[1]: https://gitlab.com/gitlab-org/gitaly/-/commit/5aae918d54bddb19743d0f2aae54948ecedac273