Render WebP attachments as image by default
This adds the .webp file extension to the list of safe image extensions
in the file type detection. It allows webp-images to be uploaded to
issues and comments which then get embeded as images instead of
attachments (see [gitlab-org/gitlab#296600](3)). It also allows uploading
of designs in the webp format. WebP is supported in all browsers that
GitLab supports except Safari <= macOS 10
(see [caniuse][1], [suppported browsers][2]).

[1]: https://caniuse.com/webp
[2]: https://docs.gitlab.com/ee/install/requirements.html#supported-web-browsers
[3]: https://gitlab.com/gitlab-org/gitlab/-/issues/296600