Fix collapsible sections start and end markers.
It took me a while to figure out how to get
collapsible sections working. Finally figured
it out by looking at the raw output and comparing
with built-in sections like `Preparing the "shell" executor`.
Only then did I notice the \e[OK at the beginning
of section_start and section_end was not mentioned
in the docs.

@gl-docsteam please review :)