Update the database when moving repos between identical storages
When any two storages are configured identically it was previously
impossible to move projects from one to the other. This is because
before trying to move the repository we check if the storages are
identical to prevent dataloss. Instead of bailing out with an exception,
now we simply update the database. This allows admins to update all
projects before fixing the configuration error.