Improve overriding stages example
For the default security jobs, the current examples
appear dated for situations where the `test` stage is
undefined. It is also lacking example names to use for
all the (currently, 4) types of includes available.

This change adds a more complete example on how to
override the stage name for every one of the included
jobs, in a copy-paste-able form.