Introduce a rendering limit for commit titles
We pass "full" commit titles through Banzai in just one place:
Projects::RefsController#logs_tree (via Gitlab::TreeSummary). It is
shown (truncated to 100 characters, with the rest in tooltips) on the
repository file browser.

Permitting unlimited-length lines to be passed through banzai is a
recipe for puma timeouts, so add a 1KiB limit to these, mirroring the
1MiB limit we added to commit descriptions.

We also clean up the omission string used (from unicode to '...') so it
renders correctly in more places.