Remove ref_path from coverage CSV path
We no longer use ref_path in the path to download
the CSV for test coverage. We were making
ref_path equal to an empty string because of
how the frontend was adding the query params.
By utilizing `mergeUrlParams`, we can get rid of
the empty string and make our path more reliable.