Wait up to 60 seconds instead of 30 seconds till PIDs should exist.
After the upgrade of puma the daemon option of puma does not exist anymore.
This was fixed with !52404.
This also fixed #299570.

But production environments I manage show that puma now requires about 30-50 seconds to start.
This will give puma enough time to start in an expected time range and not terminate to startup of gitlab.