Set has_external_wiki to NOT NULL default FALSE
This change contains migrations for `projects.has_external_wiki` to
become a `NOT NULL` column that defaults to `FALSE`.

This is possible after the PG Trigger added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/49916 that always
maintains `projects.has_external_wiki` as a boolean and never a `NULL`.

It includes a data migration to set all historical `NULL` columns to be
either `TRUE` or `FALSE`.

The data migration also fixes the historical bad data for
`has_external_wiki`:
https://gitlab.com/gitlab-org/gitlab/-/issues/273574.

https://gitlab.com/gitlab-org/gitlab/-/issues/290715.