Run migrations in reverse order in CI check
In the CI job that validates migration schema changes, the current
behavior did not account for multiple migrations in one MR, and so ran
the migrations based on the output of `git diff`.

This change fixes that, so all migrations in `post_migrate` will run
first in reverse timestamp order, and then migrations in `migrate` will
run in reverse timestamp order.