Fix changes to project that persist in spec
Applying the method in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/39996/diffs#note_402591440
to more places in the spec where `subject.project` is mutated.

This is because of the `keep_around` that is applied to projects within
this spec. For any parts of this spec where the merge request project is
mutated we need to be careful to create a new project.