Display success message when namespace is linked in Jira
Use localStorage to set and retrieve alert data,
used for success messages in Jira Connect app.

Here, we refactor SET_ERROR_MESSAGE to SET_ALERT.
This allows us to control the "global" gl-alert component
via the same mutation.

We also extend the gl-alert to include an optional
`linkUrl`, which is injected via gl-sprintf if it
exists.