Refactor: Remove requirement for local? method
For DRYness. And it removes a tiny bit of required work from the
framework doc.

I was going to write a centralized `local?` method somewhere but
discovered that all `Uploader` classes include `ObjectStorage::Concern`
which implements `file_storage?`, which is an equivalent check to these
`local?` methods.

The `local?` methods are more direct than going through
`replicator.carrierwave_uploader`, but this change is safe because we
already depend on `replicator.carrierwave_uploader` in the
`checksummable?` method.