Use the exactly 32 byte long version of db_key_base with aes-256-gcm
Treewide, usage of a truncated version of db_key_base has been common
and it works fairly well with `aes-256-cbc`. However, with
`aes-256-gcm` the key must be exactly 32 bytes long - if it's longer
it has to be truncated, but if it's shorter it has to be padded. For
this reason the such a version of the key,
`attr_encrypted_db_key_base_32`, was introduced in
777b6713bb473d2e09c8340ab9a96373fdbaae50. This changes all occurrences
of `attr_encrypted_db_key_base_truncated` being used with
`aes-256-gcm`.