Add client endpoint to filter eligible namespaces
We want to fetch the eligibility status of a namespace to purchase a
subscription from the customer's app. This way, we do not have to
replicate the eligibility logic between gitlab and the customer's app.

Add an endpoint to the client that can be used to fetch the eligibility
status.

Issue: https://gitlab.com/gitlab-org/customers-gitlab-com/-/issues/2344