Make `from` optional in the changelog API
The `from` argument in the changelog generation API is now optional.
When unspecified, GitLab will use the tag of the release that came
before the new release. This means that most users never have to specify
the `from` argument, making it easier to generate changelogs for a new
version.

This fixes
https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1542.