Email participants of on-call rotation when alerts occur
When an alert comes in through an HTTP integration or Prometheus
integration, we'll find the user(s) who are currently on-call
and notify them of the alert. Also sends out a notification
when an alert is resolved. The email sent matches the email
sent if the project has alert emails configured.