Remove adjacency checks
This removes an extra check on moving designs that fails frequently
in the presence of soft-deletes.

If we have `[a b (c) d e]` (where `c` is a deleted design, not visible
to the user), then an attempt to move `e -> (b,d)` will fail due to the
adjacency check, even though from the user's perspective, `b` and `d`
are adjacent.

As long as clients are sensible, this extra level of sanity checking
is not necessary.