Bump Elasticsearch in CI to 7.10.1
We are planning on upgrading our production environment to 7.10.2 so we
want to use the same version in CI.

There isn't, yet, a 7.10.2 published to dockerhub. Elastic also has
their own registry where they released 7.10.2 but we've had problems
with being rate limited from this registry before
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/21750 so it's best
to just go with 7.10.1 in CI for now until it's published.

I did, however, confirm in CI that 7.10.2 does work with the Elastic
docker image
https://gitlab.com/gitlab-org/gitlab/-/commit/5597a647c1092af83c717a8269bafcfe770178d0
.