Fix specs which compare UserSquash requests
With the previous change, we started to put the current system time into
a lot of RPC requests of the OperatingService. This breaks tests where
we stub out the OperatingService and expect to receive a specific
request when executing `user_squash()` as it now contains exactly that
timestamp.

Fix the issue by invoking `user_squash()` with a specific time and
changing the expected request to include that timestamp.