Introduce WordDiff parser
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/16950

We want to introduce a support for `git diff --word-diff=porcelain`
format. This format has differences from the default `git diff`.

Each line in the output can start with:
` ` - shows that data in this chunk is unchanged
`+` - shows that data in this chunk is added
`-` - shows that data in this chunk is removed
`~` - a newline character

The parser converts word-diff format to the collection of
`Gitlab::Diff::Line` object that we use for diff representation.