Update gettext-extractor and gettext-extractor-vue
We have introduced `gettext-extractor` a while in order to have proper
gettext parsing in Vue files:

https://gitlab.com/gitlab-org/gitlab-foss/-/merge_requests/19378

I took some time over the weekend to prepare the `gettext-extractor-vue`
package for `vue@3` and moved it to gitlab-org/frontend. It supports
both `vue@2` and `vue@3`, but in order to do that we had a minor change
in the API. We now pass in the template compiler from the outside.

The output should be the same, which we can see by us not needing to
regenerate the `gitlab.pot` file.