Document updated Praefect virtual storage configuration format
Praefect's virtual storage configuration format in omnibus left no
room for other virtual storage related configuration values as the
nodes were configured directly in the root of the virtual storage's
configuration object. The format was changed to address this by
moving the node configuration under the 'nodes' key. This commit
updates the documentation to use the updated format.

The change was made in a backwards compatible manner. If nodes are
configured in the root of the object, Omnibus corrects this by moving
unknown keys under the 'nodes' key.