Github Importer: Validate repository size before importing
Currently the size of the github repository is not validated before
starting the importing process. This leads to a slow feedback loop where
the failure happens much later in the process. To improve the feedback
loop and give the user better experience the validation is now done
before the import starts. If the repository to be imported is larger
than the namespace or application limits an error message expressing the
repository size and the limit size is shown.

Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/215057