Fix search functionality in Jira issues list
A recent change (still unidentified) changed the way
FilteredSearchBar sends back its `filters` object in the
`onFilter` event.

In particular, the free text filter changed from a string value
to an object with the key `filtered-search-term`. This
caused an regression in the Jira issues list as we were lacking
specs around the free text search functionality.

This fix updates the code to follow the next object structure
and adds specs to make sure it doesn't break again.