Move eslint related yarn scripts to the lint scope
We are currently overloading the `eslint` command with our custom eslint
logic. Unfortunately this causes us to run `eslint` on all js/vue files
and makes it hard to run custom eslint commands. One has to resort to
running e.g. `node_modules/.bin/eslint` directly.

In an effort to consolidate the situation we are renaming the commands
(borrowing from our rake syntax) as follows:

- `eslint` => `lint:eslint`
- `eslint-fix` => `lint:eslint:fix`
- `eslint-staged` => `lint:eslint:staged`
- `eslint-staged-fix` => `lint:eslint:staged:fix`
- `eslint-report` => `lint:eslint:report`

All but the `eslint` command have been replaced with a command that
errors out immediately and hints folks to the new command.