Move prettier related yarn scripts to the lint scope
Similar to `eslint`, this moves our `prettier` related scripts to the
`lint` scope. We keep the old scripts in place and error out
immediately, hinting the user to the new scripts.

Thanks to the introduced `scripts/frontend/execute-on-staged-files.sh`
and the fact that `prettier` itself has implemented better command line
interfaces (`--check` and `--write`) since we introduced it, we are able
to remove our home made node script in favor of those.

Now the scripts carry the following names:

 - `prettier-all` => `lint:prettier`
 - `prettier-all-save` => `lint:prettier:fix`
 - `prettier-staged` => `lint:prettier:staged`
 - `prettier-staged-save`: `lint:prettier:staged:fixed`