Use diff metadata to find out when to send diff-files-end perf mark
We were using `state.diffs.length` to find out whether the index
of diff item is the last of all diffs. But the `state.diffs.length`
doesn't reflect the total, but only the number of currently
fetched items. This change uses `diffFilesLength` instead, which is
taken from the diff metadata and reflects the total number of changed
files.