Limit number of refs shown in new pipeline page
The "Run pipeline" page does not initially filter branches or tags when
rendered, which has impact on the page performance. This change limits
the number of tags or branches rendered so the page is not overloaded.

As users add more tags and/or branches to their projects they
may use the filtered search to find what they are looking for.