Fix git access denied errors when using Spring
When allowing git disk access, we call
Gitlab::GitalyClient::StorageSettings.allow_disk_access which tracks
this using a class instance variable.

When classes are unloaded by Spring, we still keep a stale
instance of Gitlab::GitalyClient::StorageSettings in Gitlab.config.
And when self.class.disk_access_denied? is called on that instance, it
was checking the class instance variable of the stale class.

We fix this by using
Gitlab::GitalyClient::StorageSettings.disk_access_denied? so that both
of them use the same class instance variable.