Add Azure Active Directory v2 OAuth2 support
This brings in the OmniAuth provider defined in
https://github.com/RIPGlobal/omniauth-azure-activedirectory-v2. The v1
endpoints do not seem to trigger MFA, which prevents logins when MFA is
enforced via Azure's Conditional Access policy.

Note that existing identities with the legacy V1 Azure provider do not
work, and users will either have to reconnect their GitLab accounts or
admins have to enable the auto-link feature.

Note that the following permission scopes are required for the
registered Web application:

1. profile
2. openid
3. email

In addition, a redirect URI for
`/users/auth/azure_activedirectory_v2/callback` has to be added to the
Web application.

Relates to:

1. https://gitlab.com/gitlab-org/gitlab/-/issues/214390
2. https://gitlab.com/gitlab-org/gitlab/-/issues/215674