ci: fix $CI_DEPENDENCY_PROXY_GROUP_IMAGE_PREFIX for non-lowercase groups
Docker does not allow uppercase characters in the image path, but GitLab
allows them in the group name, which is used in the dependency proxy's
image prefix.

The proxy already works with lowercased group names, the only thing
needed is to make sure we pass the correct prefix to the CI and to the
web page user.