Fix error when rendering project pipeline status
`project.last_pipeline` always queries the DB for the last pipeline so
the `project.last_pipeline.present?` check does not ensure that
`project.last_pipeline` exists in the next call.

We fix this by saving the last pipeline to a local variable so that
`#present?` and `#detailed_status` are called on the same object.