Fix License Compliance widget button spacing
Use flexbox for laying out the buttons, to avoid whitespace in the
template from affecting visual spacing.

This also replaces the deprecated slot syntax with modern shorthand.

Finally, this removes a snapshot test that wasn't providing any value,
since other tests in the suite make the necessary assertions about the
`Manage licenses` and `View full report` buttons in the `ReportSection`
anyway (which is what was in the snapshot).

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/283889.