Make the variable type for the GitLab CI secret configurable
When developing the initial GitLab CI secrets support with HashiCorp
Vault as the first secret engine, we've choosen to pass the resolved
secrets to job scope as a "file type" variables.

This however quickly became a limitation as some software is unable
to work with this type of the variables nor it can be adjusted
to it. And this adds additional steps that a user needs to define in
the job to make things working.

With this change we're adding a syntax to make this configurable
and we're updating the job payload API structure.

For the full support an updated version of GitLab Runner will be
required. Support on the Runner side is being added separately.

Changelog: added
EE: true