Ignore pre-release tags when generating changelogs
When generating a changelog without specifying the `to` argument, we use
the tag of the last release. If a project has any pre-release tags, such
as GitLab itself, we would end up using such a tag. But this is
undesired, as changelogs are usually only updated for stable releases
(as is the case for GitLab for example).

To solve this, we no longer consider pre-release tags. This means that
in the above example we'd instead use the tag of the last stable
release, as one would expect.

This fixes https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1574