Support schema-deprecated arguments for GraphQL
The newer version of the graphql gem updated in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/53697 supports
marking arguments as deprecated within the schema
(https://github.com/rmosolgo/graphql-ruby/pull/3015).

Previously, deprecated arguments were only marked as such in their
description.

This change enables them to be fully marked as deprecated in the schema,
and therefore introspection queries.