Docs: Make EFS warning vendor agnostic
GitLab has a long history with supporting EFS with Gitaly, which led to
the fairly severe warning the docs currently have. This commit changes
the wording to be vendor agnostic to achieve two things:
1. Not call out a single vendor
2. Also warn about potential competing products

This change does not state EFS and others are supported at this time,
nor will networked filesystems until proven they perform.