Enable RSpec/MultipleMemoizedHelpers by default
We bring this cop into the main config, and set a lower threshold.

Some files are excluded (migration specs, which we don't want to touch),
others are fixed, and some have explicit disablements.

The threshold of 28 is chosen as it accounts for the vast majority of
our usage. We should aim to bring this down to 20, and eventually closer
to 10.

Note that at present let_it_be is not counted.