Convert GQL presenter tool to field extension
This changes the presenter instrumentation to a field extension, which
is simpler, easier to maintain, and will allow us to make use of the
interpreter execution engine in the future.

This also makes is easier to test resolver and types that make use of
these facilities.

Other changes:

In the current user TODOs type, ensure we take the type of the unwrapped
object If we don't do this, the presented class name will be used
instead (for types that use `present_using`). This results in incorrect
todo_type values.