Test fixes as a result of graphql_helpers changes
Most of these are trivial (adding `#to_a` for example, or changing
to use client side argument names). The labels_resolver_spec changes
are more invasive since this spec tested a number of impossible states
(passing arguments not declared by the resolver) - these particular
changes were approved by @acroitor as part of
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/40088.

The authorization_spec changes are made in such as way as to minimise
the impact of changes from
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/40088
from which this work was extracted.