Correct role levels required for Release CRUD
Current permission docs are incorrect when it comes to deleting a
Release object in a project.

When originally added, the permission was only given to maintainers or
higher to be able to delete a Release:
Commit 8f1e96c89bd3ffe6ae47c275df2d1e919d42c39a
from ~2018

This change corrects the permissions table and updates the API
documentation to mention the roles required when creating,
editing/updating, or deleting a Release.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/219392